rename: type 폴더
