+++ src/types/index.ts
@@ -1,2 +1,3 @@
-export * from './userInfo';
+export * from './sign-up/signUpDto';
+export * from './userInfoDto';
 export * from './filterType';
